Add profile search backend code
Adds a basic search on profiles using the first name and last name of
the profile.

It relies on the 'Profile' collection in the database having a
text index set on it. I used the pymongo command:
`db.Profile.create_index([('firstName', TEXT), ('lastName', TEXT)],
name='profile_search')` to create it. This may cause reproducibility
issues if the database is rebuilt or replaced without the corresponding
index created. To avoid this, the index should be saved in the database
schema so that when the database is rebuilt or replaced, the text index
is also created along with it.