Switch event host field to use profile ids
Previously, event host field was a field consisting of first name and
last name of the user when the event was created. However, if the user
profile changes, the event will not reflect a change in the users details,
and we will not be able to link to the profile of the host inside the
event. Switching the host field to a profile ObjectId allows us to achieve
the above.

Since searching events by host was dependent on using this first name
and last name in the host field, also add an alternative method to do
this event search by host name.